requirements.txt
4Suite-XML==1.0.2
Beaker==1.5.4
Cython==0.15.1
Django==1.8.2
Genshi==0.6
Jinja2==2.7.3
LHAPDF==6.1.5
MarkupSafe==0.23
MySQL-python==1.2.3
PyXML==0.8.4
PyYAML==3.10
SQLAlchemy==0.7.7
affinity==0.1.0
argparse==1.2.1
astroid==1.3.2
backports.ssl-match-hostname==3.4.0.2
bunch==1.0.0
certifi==14.05.14
coverage==3.5.2
cx-Oracle==5.1.1
datadiff==1.1.1
distribute==0.6.28
django-bootstrap3==6.1.0
django-debug-toolbar==1.3.2
extensions==0.4
## FIXME: could not find svn URL in dependency_links for this package:
flup==1.0.3.dev-20110405
futures==2.1.2
grin==1.2.1
h5py==2.0.0
hgsvn==0.1.9
ipython==0.12.1
iterutils==0.1.6
jsonpickle==0.4.0
keyring==0.5.1
logilab-common==0.63.2
lxml==2.3
matplotlib==1.3.1
mercurial==2.1.2
mock==0.8.0
multiprocessing==2.6.2.1
nested-dict==1.0.9
nose==1.1.2
numpy==1.8.0
ordereddict==1.1
paramiko==1.7.6
processing==0.52
py==1.4.8
py2neo==1.4.6
pyAMI==5.0.2
pyMinuit2==0.0.1
pycrypto==2.1.0
pydot==1.0.28
pyflakes==0.4.0
pyinotify==0.9.6
pylint==1.2.1
pyparsing==2.0.3
pytest==2.2.4
python-dateutil==2.4.0
pytz==2014.7
rivet==2.2.1
rootplot==2.2.1
rootpy==1.0
ruffus==2.2
scipy==0.10.0
sh==1.11
simplejson==2.1.6
six==1.8.0
sqlparse==0.1.15
stomp.py==3.1.3
storm==0.19
sympy==0.7.1
tornado==4.0.2
uncertainties==1.7
virtualenv==1.6.4
wsgiref==0.1.2
yoda==1.3.0
